LayoutTests/fast/forms/date/date-editable-components/date-editable-components-component-set-updates-color-expected.txt

PASS getComputedStyle(field).color is not "rgb(0, 0, 0)"
PASS getComputedStyle(field).color is "rgb(0, 0, 0)"
PASS successfullyParsed is true

TEST COMPLETE
